feat #84 商家账户注册（2）

+++ muyuan-member/src/main/java/com/muyuan/member/domains/vo/UserVO.java
@@ -27,12 +27,12 @@ public class UserVO {
     /**
      * 用户类型 0-个人 1-企业
      */
-    private short type;
+    private Integer type;
 
     /**
      * 店铺号
      */
-    private long shopNo;
+    private Long shopNo;
 
     /**
      * 图片
